refactor: Factor out adding blob to s3
addBlobToS3 is now a reusible DRY function. It will given a blob,
and an extension and will add to s3 and return a promise for the
resulting filename that represents the s3 key.
