+++ client/utils/index.js
@@ -1,0 +1,5 @@
+const { addBlobToS3, } = require('./s3Utils');
+
+module.exports = {
+  addBlobToS3,
+};
